store/tikv:move definition of TiKVClient to store/tikv (#22563)
